+++ Sources/MapboxMaps/Style/Style+Localization.swift
@@ -83,8 +83,8 @@ extension Style {
         let expressionAbbr = try NSRegularExpression(pattern: "\\[\"get\",\\s*\"abbr\"\\]",
                                                      options: .caseInsensitive)
 
-        if var stringExpression = String(data: try JSONEncoder().encode(symbolLayer.textField), encoding: .utf8),
-           stringExpression != "null" {
+        if case .expression(let textField) = symbolLayer.textField,
+           var stringExpression = String(data: try JSONEncoder().encode(textField), encoding: .utf8) {
             stringExpression.updateOnceExpression(replacement: replacement, regex: expressionCoalesce)
             stringExpression.updateExpression(replacement: replacement, regex: expressionAbbr)